# Ni-Ming-Gao-Bai-She-Jiao-App/artifacts/api-server/src/routes/moderation.ts
import { clerkClient, getAuth } from "@clerk/express";
import { and, count, desc, eq, isNull, sql } from "drizzle-orm";
import { Router, type IRouter, type Request, type Response } from "express";
import {
  GetModerationAccessResponse,
  ListModerationReportsQueryParams,
  ListModerationReportsResponse,
  UpdateModerationReportBody,
  UpdateModerationReportParams,
  UpdateModerationReportResponse,
} from "@workspace/api-zod";
import {
  confessionsTable,
  confessionReportsTable,
  announcementsTable,
  insertAnnouncementSchema,
  bannedUsersTable, // 順便把黑名單表加進來
  db,
} from "@workspace/db";
import { logger } from "../lib/logger";
import { publishConfessionToInstagram } from "../lib/instagram";

const router: IRouter = Router();
type ModeratorCheck = "allowed" | "denied" | "unavailable";

export async function checkModerator(userId: string): Promise<ModeratorCheck> {
  const allowedEmails = new Set(
    (process.env.CONFESSION_MODERATOR_EMAILS ?? "")
      .split(",")
      .map((email) => email.trim().toLowerCase())
      .filter(Boolean),
  );
  if (allowedEmails.size === 0) return "unavailable";

  try {
    const user = await clerkClient.users.getUser(userId);
    const isAllowed = user.emailAddresses.some(
      (address) =>
        address.verification?.status === "verified" &&
        allowedEmails.has(address.emailAddress.toLowerCase()),
    );
    return isAllowed ? "allowed" : "denied";
  } catch {
    logger.error("Unable to verify moderator identity with Clerk");
    return "unavailable";
  }
}

function reportSelection() {
  return {
    id: confessionReportsTable.id,
    confessionId: confessionReportsTable.confessionId,
    reason: confessionReportsTable.reason,
    details: confessionReportsTable.details,
    status: confessionReportsTable.status,
    reportedAt: confessionReportsTable.createdAt,
    confessionContent: confessionsTable.content,       // 維持原本的 confessionContent
    confessionCategory: confessionsTable.category,     // 維持原本的 confessionCategory
    confessionCreatedAt: confessionsTable.createdAt,   // 維持原本的 confessionCreatedAt
  };
}

async function hasModeratorAccess(
  req: Request,
  res: Response,
): Promise<boolean> {
  const userId = getAuth(req).userId;
  if (!userId) {
    res.status(401).json({ error: "Sign-in required" });
    return false;
  }

  const check = await checkModerator(userId);
  if (check === "unavailable") {
    res.status(503).json({ error: "Moderator access is temporarily unavailable" });
    return false;
  }
  if (check === "denied") {
    res.status(403).json({ error: "Moderator access required" });
    return false;
  }
  return true;
}

router.get("/moderation/access", async (req, res): Promise<void> => {
  const userId = getAuth(req).userId;
  if (!userId) {
    res.status(401).json({ error: "Sign-in required" });
    return;
  }

  const check = await checkModerator(userId);
  if (check === "unavailable") {
    res.status(503).json({ error: "Moderator access is temporarily unavailable" });
    return;
  }

  let newReports = 0;
  if (check === "allowed") {
    const [{ value }] = await db
      .select({ value: count() })
      .from(confessionReportsTable)
      .where(eq(confessionReportsTable.status, "new"));
    newReports = value;
  }

  res.json(
    GetModerationAccessResponse.parse({
      allowed: check === "allowed",
      newReports,
    }),
  );
});

router.get("/moderation/reports", async (req, res): Promise<void> => {
  if (!(await hasModeratorAccess(req, res))) return;

  const parsed = ListModerationReportsQueryParams.safeParse(req.query);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const statusFilter =
    parsed.data.status === "all"
      ? undefined
      : eq(confessionReportsTable.status, parsed.data.status);

  const reports = await db
    .select(reportSelection())
    .from(confessionReportsTable)
    .innerJoin(
      confessionsTable,
      eq(confessionReportsTable.confessionId, confessionsTable.id),
    )
    .where(statusFilter)
    .orderBy(desc(confessionReportsTable.createdAt), desc(confessionReportsTable.id));

  // 1. 先透過 Zod 驗證確保資料結構正確（避免 500 錯誤）
  const validatedReports = ListModerationReportsResponse.parse(reports);

  // 2. 映射欄位，同時提供前端需要的 content 與 category
  const responseData = validatedReports.map((report) => ({
    ...report,
    content: report.confessionContent,
    category: report.confessionCategory,
  }));

  res.json(responseData);
});

router.patch("/moderation/reports/:reportId", async (req, res): Promise<void> => {
  if (!(await hasModeratorAccess(req, res))) return;

  const params = UpdateModerationReportParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }
  // 改用更彈性的方式驗證 status，允許 resolved 與 dismissed
const body = req.body as { status?: string };
const allowedStatuses = ["new", "reviewed", "resolved", "dismissed"];

if (!body.status || !allowedStatuses.includes(body.status)) {
  res.status(400).json({ error: "Invalid status value" });
  return;
}

const newStatus = body.status;

  const [updated] = await db
    .update(confessionReportsTable)
    .set({ status: newStatus }) // 這裡改成 newStatus
    .where(eq(confessionReportsTable.id, params.data.reportId))
    .returning({ id: confessionReportsTable.id });

  // 假設你原本是從資料庫撈出 [report]，例如：
  const [report] = await db
    .select(reportSelection())
    .from(confessionReportsTable)
    .innerJoin(
      confessionsTable,
      eq(confessionReportsTable.confessionId, confessionsTable.id),
    )
    .where(eq(confessionReportsTable.id, params.data.reportId))
    .limit(1);

  if (!report) {
    res.status(404).json({ error: "Report not found" });
    return;
  }

  // 👇 請在這裡加上全方位的欄位對應，再回傳給前端
  const responseData = {
    ...report,
    content: report.confessionContent,
    confessionContent: report.confessionContent,
    postContent: report.confessionContent,
    category: report.confessionCategory,
    confessionCategory: report.confessionCategory,
  };

  res.json(responseData);
});

// 1. Delete confession post by ID
// Delete confession post by ID
router.delete("/confessions/:id", async (req: Request, res: Response) => {
  const auth = getAuth(req);
  if (!auth.userId) {
    return res.status(401).json({ message: "Unauthorized" });
  }

  const postId = Number(req.params.id);
  if (isNaN(postId)) {
    return res.status(400).json({ message: "Invalid post ID" });
  }

  try {
    // Delete target confession from database
    await db.delete(confessionsTable).where(eq(confessionsTable.id, postId));

    // Mark corresponding reports as resolved
    await db
      .update(confessionReportsTable)
      .set({ status: "resolved" })
      .where(eq(confessionReportsTable.confessionId, postId));

    return res.json({ success: true, message: "Post deleted successfully" });
  } catch (error) {
    logger.error({ error }, "Failed to delete post");
    return res.status(500).json({ message: "Failed to delete post" });
  }
});

const handleDeleteConfession = async (req: Request, res: Response) => {
  const auth = getAuth(req);
  if (!auth.userId) {
    return res.status(401).json({ message: "Unauthorized: Please log in." });
  }

  const postId = Number(req.params.id);
  if (isNaN(postId)) {
    return res.status(400).json({ message: "Invalid post ID" });
  }

  try {
    await db.transaction(async (tx) => {
      // 1. 先清理举报记录
      await tx
        .delete(confessionReportsTable)
        .where(eq(confessionReportsTable.confessionId, postId));

      // 2. 如果你的 schema 里有评论表或点赞表，请取消对应注释并添加：
      // await tx.delete(confessionCommentsTable).where(eq(confessionCommentsTable.confessionId, postId));
      // await tx.delete(confessionLikesTable).where(eq(confessionLikesTable.confessionId, postId));

      // 3. 最后删除告白贴文主记录
      await tx
        .delete(confessionsTable)
        .where(eq(confessionsTable.id, postId));
    });

    return res.json({ success: true, message: "Post deleted successfully" });
  } catch (error: any) {
    logger.error({ error, postId }, "Failed to delete post inside transaction");
    return res.status(500).json({ 
      message: error?.message || error?.detail || "Database cascade delete failed." 
    });
  }
};

// 注册路由接口
router.delete("/confessions/:id", handleDeleteConfession);
router.delete("/posts/:id", handleDeleteConfession);

router.delete("/confessions/:id", handleDeleteConfession);
router.delete("/posts/:id", handleDeleteConfession);

router.delete("/confessions/:id", handleDeleteConfession);
router.delete("/posts/:id", handleDeleteConfession);

router.delete("/confessions/:id", handleDeleteConfession);
router.delete("/posts/:id", handleDeleteConfession);

// Register routes
router.delete("/confessions/:id", handleDeleteConfession);
router.delete("/posts/:id", handleDeleteConfession);

// Register routes for both endpoints
router.delete("/confessions/:id", handleDeleteConfession);
router.delete("/posts/:id", handleDeleteConfession);

// Ban user via local database blacklist (Bypassing Clerk payment restriction)
router.post("/users/ban", async (req: Request, res: Response) => {
  const auth = getAuth(req);
  if (!auth.userId) {
    return res.status(401).json({ message: "Unauthorized" });
  }

  const { confessionId } = req.body;
  if (!confessionId) {
    return res.status(400).json({ message: "Missing confessionId in request body" });
  }

  try {
    // 0. 自動確保雲端資料庫存在正確的 banned_users 表格與主鍵
    await db.execute(sql`
      CREATE TABLE IF NOT EXISTS "banned_users" (
          "user_id" text PRIMARY KEY NOT NULL,
          "banned_at" timestamp DEFAULT now() NOT NULL
      );
    `);

    // 1. 透過前端傳來的 confessionId 去資料庫找出這篇貼文的作者 (authorId)
    const post = await db
      .select()
      .from(confessionsTable)
      .where(eq(confessionsTable.id, Number(confessionId)))
      .limit(1);

    if (post.length === 0) {
      return res.status(404).json({ message: "Confession post not found in database" });
    }

    const targetUserId = post[0].authorId;
    if (!targetUserId) {
      return res.status(404).json({ message: "Cannot ban: Author ID is missing in the database record." });
    }

    // 2. 將該使用者的 ID 寫入黑名單
    await db.insert(bannedUsersTable)
      .values({ userId: targetUserId })
      .onConflictDoNothing({ target: bannedUsersTable.userId });

    return res.json({ success: true, message: "User added to local blacklist successfully" });
  } catch (error: any) {
    logger.error({ error }, "Failed to ban user");
    return res.status(500).json({ message: error?.message || "Failed to ban user due to server error" });
  }
});

// ==========================================
// 新增：發布系統公告 API
// ==========================================
// 獲取所有系統公告
// 獲取所有系統公告
// 發布系統公告 API
router.get("/announcements", async (req, res) => {
  try {
    const list = await db
      .select()
      .from(announcementsTable)
      .orderBy(desc(announcementsTable.createdAt));
    res.json(list);
  } catch (error) {
    console.error("Failed to fetch announcements:", error);
    res.status(500).json({ error: "Failed to fetch announcements" });
  }
});

// 發布系統公告 API
router.post("/announcements", async (req, res) => {
  try {
    const parsed = insertAnnouncementSchema.parse(req.body);
    const [newAnnouncement] = await db
      .insert(announcementsTable)
      .values(parsed)
      .returning();
    
    res.status(201).json(newAnnouncement);
  } catch (error) {
    console.error("Failed to create announcement:", error);
    res.status(400).json({ error: "Failed to create announcement" });
  }
});

// 刪除系統公告 API
// 刪除系統公告 API（加上管理員信箱驗證）
router.delete("/announcements/:id", async (req, res) => {
  try {
    const userEmail = req.headers["x-user-email"] || req.body?.email;
    const moderatorEmails = process.env.CONFESSION_MODERATOR_EMAILS || "";
    
    // 🔍 加上這行來印出除錯訊息，查看雲端實際收到的內容
    console.log("【Access Control Check】Received email address:", userEmail, "| Allowed administrator mailbox:", moderatorEmails);

    if (!userEmail || !moderatorEmails.includes(String(userEmail))) {
      return res.status(403).json({ error: "Unauthorized: You are not a moderator" });
    }

    const id = parseInt(req.params.id);
    await db
      .delete(announcementsTable)
      .where(eq(announcementsTable.id, id));
    
    return res.json({ success: true });
  } catch (error) {
    console.error("Failed to delete announcement:", error);
    return res.status(500).json({ error: "Failed to delete announcement" });
  }
});

router.post("/confessions/:id/instagram", async (req: Request, res: Response): Promise<void> => {
  if (!(await hasModeratorAccess(req, res))) return;

  const id = Number(req.params.id);
  if (!Number.isInteger(id) || id <= 0) {
    res.status(400).json({ error: "Invalid confession ID" });
    return;
  }

  const [claimed] = await db
    .update(confessionsTable)
    .set({ igPostedAt: new Date() })
    .where(and(eq(confessionsTable.id, id), isNull(confessionsTable.igPostedAt)))
    .returning({ id: confessionsTable.id });

  if (!claimed) {
    const [existing] = await db
      .select({ id: confessionsTable.id })
      .from(confessionsTable)
      .where(eq(confessionsTable.id, id))
      .limit(1);
    res
      .status(existing ? 409 : 404)
      .json({ error: existing ? "Already posted to Instagram" : "Confession not found" });
    return;
  }

  let mediaId: string;
  try {
    mediaId = await publishConfessionToInstagram(id);
  } catch (error) {
    logger.error({ err: error, confessionId: id }, "Failed to publish to Instagram");
    await db
      .update(confessionsTable)
      .set({ igPostedAt: null })
      .where(eq(confessionsTable.id, id));
    res.status(502).json({ error: "Failed to publish to Instagram" });
    return;
  }

  await db
    .update(confessionsTable)
    .set({ igMediaId: mediaId })
    .where(eq(confessionsTable.id, id));
  res.json({ success: true, igMediaId: mediaId });
});

export default router;