# Ni-Ming-Gao-Bai-She-Jiao-App/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /confessions/{id}:
    get:
      operationId: getConfession
      tags: [confessions]
      summary: Get one confession
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Confession
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Confession"
        "404":
          description: Confession not found
...
  /moderation/access:
    get:
      operationId: getModerationAccess
      tags: [moderation]
      summary: Check whether the signed-in user can moderate reports
      responses:
        "200":
          description: Moderation access status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModerationAccess"
        "401":
          description: Sign-in required
        "503":
          description: Moderator access is not configured or unavailable
  /moderation/reports:
    get:
      operationId: listModerationReports
      tags: [moderation]
      summary: List confession reports for moderators
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [new, reviewed, all]
            default: new
      responses:
        "200":
          description: Moderation reports
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModerationReportList"
        "401":
          description: Sign-in required
        "403":
          description: Moderator access required
        "503":
          description: Moderator access is not configured or unavailable
  /moderation/reports/{reportId}:
    patch:
      operationId: updateModerationReport
      tags: [moderation]
      summary: Update a report's review status
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModerationReportUpdate"
      responses:
        "200":
          description: Updated report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModerationReport"
        "400":
          description: Invalid report or status
        "401":
          description: Sign-in required
        "403":
          description: Moderator access required
        "404":
          description: Report not found
        "503":
          description: Moderator access is not configured or unavailable
components:
  schemas:
...
    Confession:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        category:
          type: string
          enum: [love, friendship, family, school, work, life]
        createdAt:
          type: string
          format: date-time
        likes:
          type: integer
        commentsCount:
          type: integer
        likedByMe:
          type: boolean
      required: [id, content, category, createdAt, likes, commentsCount, likedByMe]
...
    ModerationAccess:
      type: object
      properties:
        allowed:
          type: boolean
        newReports:
          type: integer
      required: [allowed, newReports]
    ModerationReport:
      type: object
      properties:
        id:
          type: integer
        confessionId:
          type: integer
        reason:
          type: string
          enum: [harassment, spam, harmful, other]
        details:
          type: ["string", "null"]
        status:
          type: string
          enum: [new, reviewed]
        reportedAt:
          type: string
          format: date-time
        confessionContent:
          type: string
        confessionCategory:
          type: string
          enum: [love, friendship, family, school, work, life]
        confessionCreatedAt:
          type: string
          format: date-time
      required:
        - id
        - confessionId
        - reason
        - details
        - status
        - reportedAt
        - confessionContent
        - confessionCategory
        - confessionCreatedAt
    ModerationReportList:
      type: array
      items:
        $ref: "#/components/schemas/ModerationReport"
    ModerationReportUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [new, reviewed]
      required: [status]
...